TRIX-293 and other minor fixes and cleanups

See merge request clustervision/luna2-daemon!367

+++ daemon/base/network.py
@@ -162,6 +162,15 @@ class Network():
                 status=False
                 ret_msg = 'Invalid request: Not enough details provided. network/subnet in CIDR notation expected'
                 return status, ret_msg
+
+            where = f" WHERE (`name`!='{name}' AND `name`!='{data['name']}')"
+            where += f" AND `network`='{data['network']}' AND `subnet`='{data['subnet']}'"
+            claship = Database().get_record(None, 'network', where)
+            if claship:
+                status=False
+                ret_msg = f"Invalid request: Clashing network/subnet with existing network {claship[0]['name']}"
+                return status, ret_msg
+
             if 'zone' in data:
                 if (data['zone'] != "external") and (data['zone'] != "internal"):
                     status=False
@@ -297,16 +306,25 @@ class Network():
         status=False
         network = Database().get_record(None, 'network', f' WHERE `name` = "{name}"')
         if network:
-            Database().delete_row('network', [{"column": "name", "value": name}])
-            data = {}
-            data['shared'] = ""
-            row = Helper().make_rows(data)
-            where = [{"column": "shared", "value": name}]
-            Database().update('network', row, where)
-            Service().queue('dns','restart')
-            Service().queue('dhcp','restart')
-            response = 'Network removed'
-            status=True
+            controller = Database().get_record_join(
+                ['controller.*'],
+                ['ipaddress.tablerefid=controller.id','network.id=ipaddress.networkid'],
+                ['tableref="controller"','controller.hostname="controller"',f"network.name='{name}'"]
+            )
+            if not controller:
+                Database().delete_row('network', [{"column": "name", "value": name}])
+                data = {}
+                data['shared'] = ""
+                row = Helper().make_rows(data)
+                where = [{"column": "shared", "value": name}]
+                Database().update('network', row, where)
+                Service().queue('dns','restart')
+                Service().queue('dhcp','restart')
+                response = 'Network removed'
+                status=True
+            else:
+                response = f'Network {name} cannot be removed because it is in use by controller'
+                status=False
         else:
             response = f'Network {name} not present in database'
             status=False
@@ -406,7 +424,6 @@ class Network():
                     device_name = ""
                 response = {'config': {'network': {name: {'taken': taken} } } }
                 status=True
-                self.logger.info(response)
             else:
                 response = 'All IP Address are free on Network {name}. None is Taken.'
                 status=False
